refactor: image module
